RS_Invoice

+++ LoginForm/LoginForm/DataSet/RS_InvoiceDetails.cs
@@ -29,7 +29,7 @@ namespace LoginForm.DataSet
         public string CCCNNO { get; set; }
         public string CountryofOrigin { get; set; }
         public string ArticleDescription { get; set; }
-        public Nullable<int> DeliveryNumber { get; set; }
+        public Nullable<decimal> DeliveryNumber { get; set; }
         public Nullable<int> DeliveryItemNumber { get; set; }
     
         public virtual RS_Invoice RS_Invoice { get; set; }
